+++ CHANGELOG.md
@@ -53,8 +53,10 @@ but may switch to [CalVer Versioning](https://calver.org/) in the future.
 #### Added
 
 - Implemented basic syntax errors fault tolerance
+<!--  -->
 - Implemented position independent cli arguments
 - Implemented cli flags colorization
+- Implemented `:` cli arguments separator: `--color:auto`, `-o:out`
 
 #### Changed